+++ src/D3D/VertexBuffer.h
@@ -11,7 +11,7 @@ using namespace DirectX;
 class VertexBuffer {
 public:
 	VertexBuffer() = default;
-	VertexBuffer(std::vector<XMFLOAT3> vertexData, size_t vertexDataSize, size_t vertexSize);
+	VertexBuffer(std::vector<XMFLOAT3> vertexData, UINT vertexDataSize, UINT vertexSize);
 
 	D3D12_VERTEX_BUFFER_VIEW passVertexDataToGPU(DXContext& context, ID3D12GraphicsCommandList6* cmdList);
 
@@ -24,7 +24,7 @@ private:
 	ComPointer<ID3D12Resource1> uploadBuffer;
 	ComPointer<ID3D12Resource1> vertexBuffer;
 
-	size_t vertexDataSize;
-	size_t vertexSize;
+	UINT vertexDataSize;
+	UINT vertexSize;
 	std::vector<XMFLOAT3> vertexData;
 };
